integrated_app/requirements.txt
Flask==3.1.2
Flask-CORS==4.0.0
gunicorn==22.0.0
python-dotenv
psycopg2-binary
bcrypt
pycryptodome
openpyxl
pyodbc
lunarcalendar==0.0.9
ephem==4.2
python-dateutil==2.9.0.post0
pytz==2025.2

# 프론트엔드 빌드 도구 (선택사항)
minify-html
cssmin
jsmin
Pillow
